\name{Rfiglet-package}
\alias{Rfiglet-package}
\docType{package}

\title{R Bindings for figlet}

\description{
  R bindings for figlet. figlet, which stands for "Frank, Ian, and 
  Glenn's LETters", is a command line utility 
}
\details{
\tabular{ll}{
  Package: \tab Rfiglet\cr
  Type: \tab Package\cr
  License: \tab 2-clause BSD\cr
}
}
\references{
  figlet
  \url{to/do/FIXME}
}
\author{
  Drew Schmidt \email{wrathematics AT gmail.com}
}
\keyword{Package}
