% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/figlet.R
\name{figlet}
\alias{figlet}
\title{FIGlet}
\usage{
figlet(text, font = "standard")
}
\arguments{
\item{text}{Text to make bigger}

\item{font}{Name of the font}
}
\description{
FIGlet
}
\examples{
message(figlet("FIGlet"))
}

