% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extra.R
\name{figlet_download_fonts}
\alias{figlet_download_fonts}
\title{Download more fonts}
\usage{
figlet_download_fonts(dest = NULL, verbose = TRUE)
}
\arguments{
\item{dest}{Directory to use. We use the environment variable
\code{RFIGLET_FONT_DIR} if defined, then fall back on
\code{tools::R_user_dir("rfiglet", "data")}}

\item{verbose}{Be verbose?}
}
\description{
Download a large set of fonts. These will be available to \link{figlet}
and \link{figlet_font} by name once downloaded.
}
\examples{
if (interactive()) {
  path <- tempfile()
  rfiglet::figlet_download_fonts(path)
  rfiglet::figlet("hello", file.path(path, "emboss.tlf"))
}
}
