% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/font.R
\name{figlet_font}
\alias{figlet_font}
\alias{figlet_font_list}
\title{Get figlet font}
\usage{
figlet_font(font, verbose = FALSE)

figlet_font_list()
}
\arguments{
\item{font}{Path or name of the font to load}

\item{verbose}{Logical, indicating if we should be verbose during font
load.}
}
\value{
A \code{figlet_font} object for use with \link{figlet}
}
\description{
Get a figlet font, returning a \code{figlet_font} object that can be
used with \link{figlet}.
}
\details{
This function tries to do the right thing with loading fonts:
\enumerate{
\item if \code{font} is a \code{figlet_font} already, return it
\item if \code{font} is a known \code{figlet_font} in the internal font registry,
then return it (faster than reading it again)
\item if \code{font} is a filename, then load the font and add it to the registry
}
}
\examples{
# Load a font with a full path
rfiglet::figlet_font(system.file("fonts/standard.flf", package = "rfiglet"))

# Load a previously seen font with a name
rfiglet::figlet_font("standard")

# List known fonts
rfiglet::figlet_font_list()
}
