% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/fonts.R
\name{figlet_fonts}
\alias{figlet_fonts}
\title{List or search figlet fonts}
\usage{
figlet_fonts(height = NULL)
}
\arguments{
\item{height}{Optional height - can be length 1 (exact height) or
length 2, to find fonts within a height range.}
}
\description{
List or search figlet fonts
}
\examples{
length(figlet_fonts())
figlet_fonts(height=10)
figlet_fonts(height=c(15, Inf))
}

