% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figlet.r
\name{print-figlet}
\alias{print-figlet}
\alias{print.figlet}
\title{Printing}
\usage{
\method{print}{figlet}(x, ...)
}
\arguments{
\item{x}{\code{figlet} object}

\item{...}{unused}
}
\description{
Printing for \code{figlet()}
}

