% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rfiglet-package.R
\docType{package}
\name{Rfiglet-package}
\alias{Rfiglet-package}
\title{R Bindings for figlet}
\description{
R bindings for figlet. figlet, which stands for "Frank, Ian, and Glenn's
LETters", is a command line utility
}
\details{
\tabular{ll}{
  Package: \tab Rfiglet\cr 
  Type: \tab Package\cr 
  License: \tab 2-clause BSD\cr
}
}
\author{
Drew Schmidt \email{wrathematics AT gmail.com}
}
\references{
figlet \url{http://www.figlet.org/}
}
\keyword{Package}

