% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figlet.R
\name{figlet}
\alias{figlet}
\title{FIGlet}
\usage{
figlet(
  text,
  font = "standard",
  width = getOption("width", 80),
  justify = "left",
  absolute = FALSE,
  strip = TRUE
)
}
\arguments{
\item{text}{Text to make bigger}

\item{font}{Name of font, path to font, or \code{figlet_font} object;
see \link{figlet_font} for details}

\item{width}{Width to use when justifying and breaking lines}

\item{justify}{Text justification to use in rendering ("left",
"centre", "right")}

\item{absolute}{Logical, indicating if alignment is absolute
(relative to \code{width}). Has an effect when justify is \code{right} or
\code{centre} only.}

\item{strip}{Logical, indicating if whitespace (trailing, plus
leading/trailing empty lines) should be removed.}
}
\value{
An object of class \code{figlet_text} which is a character
vector with a handy print method and attributes \code{font} and
\code{text}
}
\description{
FIGlet
}
\examples{
rfiglet::figlet("FIGlet")
}
