% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figlet.r
\name{figlet}
\alias{figlet}
\title{Printing via figlet}
\usage{
figlet(message, font = "standard", smush = TRUE,
  respect.linebreaks = FALSE)
}
\arguments{
\item{message}{String meant to be printed via figlet.}

\item{font}{figlet font to use.}

\item{smush}{logical; whether or not the kerning should be "extreme" (TRUE)
or "normal" (FALSE).}

\item{respect.linebreaks}{logical; if \code{TRUE}, then linebreaks will be
printed as linebreaks, and otherwise as spaces.}
}
\value{
The return is an S3-\code{figlet} object.
}
\description{
Printing via figlet.  See \code{help(figletfonts)} for a list of fonts.
}
\examples{
\dontrun{
library(Rfiglet)
figlet("abacabb")
}

}

